+++ knytt/objects/GDKnyttObjectFactory.cs
@@ -73,6 +73,8 @@ public static class GDKnyttObjectFactory
         ObjectLookup[new KnyttPoint(14, 6)] =  "PassiveWalker";
         ObjectLookup[new KnyttPoint(14, 7)] =  "PassiveWalker";
         ObjectLookup[new KnyttPoint(14, 8)] =  "PassiveWalker";
+        ObjectLookup[new KnyttPoint(15, 6)] =  "ProximityBlock";
+        ObjectLookup[new KnyttPoint(15, 7)] =  "ProximityBlock";
         ObjectLookup[new KnyttPoint(15, 27)] =  "Door";
         ObjectLookup[new KnyttPoint(15, 28)] =  "Door";
         ObjectLookup[new KnyttPoint(15, 29)] =  "Door";