+++ Changelog.md
@@ -9,6 +9,7 @@ All notable changes to YKnytt will be documented here.
 - Warp object
 - No Wall object
 - Ghost Marker
+- Proximity Blocks
 
 ## Fixed
 - Rain drops now spawn down a bit (They were colliding with the tile above)